legaily/legaily-doc-ai/server/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pdfplumber
pillow
cohere
dotenv
torch
transformers